[Decomp][Spell] Cohesion split: extract area/raid targeting to SpellAreaTargets.cpp
