+++ src/game/WorldHandlers/Spell.cpp
@@ -81,56 +81,6 @@
 #include "LuaEngine.h"
 #endif /* ENABLE_ELUNA */
 
-class PrioritizeManaUnitWraper
-{
-    public:
-        explicit PrioritizeManaUnitWraper(Unit* unit) : i_unit(unit)
-        {
-            uint32 maxmana = unit->GetMaxPower(POWER_MANA);
-            i_percent = maxmana ? unit->GetPower(POWER_MANA) * 100 / maxmana : 101;
-        }
-        Unit* getUnit() const { return i_unit; }
-        uint32 getPercent() const { return i_percent; }
-    private:
-        Unit* i_unit;
-        uint32 i_percent;
-};
-
-struct PrioritizeMana
-{
-    int operator()(PrioritizeManaUnitWraper const& x, PrioritizeManaUnitWraper const& y) const
-    {
-        return x.getPercent() > y.getPercent();
-    }
-};
-
-typedef std::priority_queue<PrioritizeManaUnitWraper, std::vector<PrioritizeManaUnitWraper>, PrioritizeMana> PrioritizeManaUnitQueue;
-
-class PrioritizeHealthUnitWraper
-{
-    public:
-        explicit PrioritizeHealthUnitWraper(Unit* unit) : i_unit(unit)
-        {
-            uint32 maxhealth = unit->GetMaxHealth();
-            i_percent = maxhealth ? unit->GetHealth() * 100 / maxhealth : 0;
-        }
-        Unit* getUnit() const { return i_unit; }
-        uint32 getPercent() const { return i_percent; }
-    private:
-        Unit* i_unit;
-        uint32 i_percent;
-};
-
-struct PrioritizeHealth
-{
-    int operator()(PrioritizeHealthUnitWraper const& x, PrioritizeHealthUnitWraper const& y) const
-    {
-        return x.getPercent() > y.getPercent();
-    }
-};
-
-typedef std::priority_queue<PrioritizeHealthUnitWraper, std::vector<PrioritizeHealthUnitWraper>, PrioritizeHealth> PrioritizeHealthUnitQueue;
-
 /**
  * @brief Checks whether a spell matches the quest tame spell pattern.
  *
@@ -1172,125 +1122,6 @@ SpellCastResult Spell::CanOpenLock(SpellEffectIndex effIndex, uint32 lockId, Ski
     }
 
     return SPELL_CAST_OK;
-}
-
-/**
- * Fill target list by units around (x,y) points at radius distance
-
- * @param targetUnitMap        Reference to target list that filled by function
- * @param radius               Radius around (x,y) for target search
- * @param pushType             Additional rules for target area selection (in front, angle, etc)
- * @param spellTargets         Additional rules for target selection base at hostile/friendly state to original spell caster
- * @param originalCaster       If provided set alternative original caster, if =NULL then used Spell::GetAffectiveObject() return
- */
-void Spell::FillAreaTargets(UnitList& targetUnitMap, float radius, SpellNotifyPushType pushType, SpellTargets spellTargets, WorldObject* originalCaster /*=NULL*/)
-{
-    MaNGOS::SpellNotifierCreatureAndPlayer notifier(*this, targetUnitMap, radius, pushType, spellTargets, originalCaster);
-    Cell::VisitAllObjects(notifier.GetCenterX(), notifier.GetCenterY(), m_caster->GetMap(), notifier, radius);
-}
-
-/**
- * @brief Fills a target list with party or raid members around a reference unit.
- *
- * @param targetUnitMap The target list being populated.
- * @param member The reference member.
- * @param radius The search radius.
- * @param raid True to include the whole raid; false to limit to the subgroup.
- * @param withPets True to include pets.
- * @param withcaster True to include the caster when applicable.
- */
-void Spell::FillRaidOrPartyTargets(UnitList& targetUnitMap, Unit* member, Unit* center, float radius, bool raid, bool withPets, bool withcaster)
-{
-    Player* pMember = member->GetCharmerOrOwnerPlayerOrPlayerItself();
-    Group* pGroup = pMember ? pMember->GetGroup() : NULL;
-
-    if (pGroup)
-    {
-        uint8 subgroup = pMember->GetSubGroup();
-
-        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
-        {
-            Player* Target = itr->getSource();
-
-            // IsHostileTo check duel and controlled by enemy
-            if (Target && (raid || subgroup == Target->GetSubGroup())
-                && !m_caster->IsHostileTo(Target))
-            {
-                if ((Target == center || center->IsWithinDistInMap(Target, radius)) &&
-                        (withcaster || Target != m_caster))
-                {
-                    targetUnitMap.push_back(Target);
-                }
-
-                if (withPets)
-                    if (Pet* pet = Target->GetPet())
-                        if ((pet == center || center->IsWithinDistInMap(pet, radius)) &&
-                                (withcaster || pet != m_caster))
-                        {
-                            targetUnitMap.push_back(pet);
-                        }
-            }
-        }
-    }
-    else
-    {
-        Unit* ownerOrSelf = pMember ? pMember : member->GetCharmerOrOwnerOrSelf();
-        if ((ownerOrSelf == center || center->IsWithinDistInMap(ownerOrSelf, radius)) &&
-                (withcaster || ownerOrSelf != m_caster))
-        {
-            targetUnitMap.push_back(ownerOrSelf);
-        }
-
-        if (withPets)
-        {
-            if (Pet* pet = ownerOrSelf->GetPet())
-            {
-                if ((pet == center || center->IsWithinDistInMap(pet, radius)) &&
-                        (withcaster || pet != m_caster))
-                {
-                    targetUnitMap.push_back(pet);
-                }
-            }
-        }
-    }
-}
-
-void Spell::FillRaidOrPartyManaPriorityTargets(UnitList& targetUnitMap, Unit* member, Unit* center, float radius, uint32 count, bool raid, bool withPets, bool withCaster)
-{
-    FillRaidOrPartyTargets(targetUnitMap, member, center, radius, raid, withPets, withCaster);
-
-    PrioritizeManaUnitQueue manaUsers;
-    for (UnitList::const_iterator itr = targetUnitMap.begin(); itr != targetUnitMap.end(); ++itr)
-        if ((*itr)->GetPowerType() == POWER_MANA && !(*itr)->IsDead())
-        {
-            manaUsers.push(PrioritizeManaUnitWraper(*itr));
-        }
-
-    targetUnitMap.clear();
-    while (!manaUsers.empty() && targetUnitMap.size() < count)
-    {
-        targetUnitMap.push_back(manaUsers.top().getUnit());
-        manaUsers.pop();
-    }
-}
-
-void Spell::FillRaidOrPartyHealthPriorityTargets(UnitList& targetUnitMap, Unit* member, Unit* center, float radius, uint32 count, bool raid, bool withPets, bool withCaster)
-{
-    FillRaidOrPartyTargets(targetUnitMap, member, center, radius, raid, withPets, withCaster);
-
-    PrioritizeHealthUnitQueue healthQueue;
-    for (UnitList::const_iterator itr = targetUnitMap.begin(); itr != targetUnitMap.end(); ++itr)
-        if (!(*itr)->IsDead())
-        {
-            healthQueue.push(PrioritizeHealthUnitWraper(*itr));
-        }
-
-    targetUnitMap.clear();
-    while (!healthQueue.empty() && targetUnitMap.size() < count)
-    {
-        targetUnitMap.push_back(healthQueue.top().getUnit());
-        healthQueue.pop();
-    }
 }
 
 /**
