+++ src/game/WorldHandlers/SpellAreaTargets.cpp
@@ -1,0 +1,251 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file Spell.cpp
+ * @brief Spell casting and effect implementation
+ *
+ * This file implements the Spell class which handles spell casting:
+ * - Spell validation and casting requirements
+ * - Spell effect execution (damage, healing, summon, etc.)
+ * - Spell targeting and area effects
+ * - Spell cooldowns and resource costs
+ * - Spell interruption and pushback
+ * - Spell aura application
+ * - Spell hit/miss calculations
+ *
+ * Spells are the primary combat mechanic in WoW, encompassing
+ * abilities, talents, and item effects.
+ *
+ * @see Spell for the spell class
+ * @see SpellAura for spell auras
+ * @see SpellMgr for spell management
+ */
+
+#include "Spell.h"
+#include "Database/DatabaseEnv.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "Opcodes.h"
+#include "Log.h"
+#include "UpdateMask.h"
+#include "World.h"
+#include "ObjectMgr.h"
+#include "SpellMgr.h"
+#include "Player.h"
+#include "Pet.h"
+#include "Unit.h"
+#include "DynamicObject.h"
+#include "Group.h"
+#include "UpdateData.h"
+#include "MapManager.h"
+#include "ObjectAccessor.h"
+#include "CellImpl.h"
+#include "Policies/Singleton.h"
+#include "SharedDefines.h"
+#include "LootMgr.h"
+#include "VMapFactory.h"
+#include "BattleGround/BattleGround.h"
+#include "Util.h"
+#include "Chat.h"
+#include "DB2Stores.h"
+#include "SQLStorages.h"
+#include "Vehicle.h"
+#include "TemporarySummon.h"
+#include "SQLStorages.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+class PrioritizeManaUnitWraper
+{
+    public:
+        explicit PrioritizeManaUnitWraper(Unit* unit) : i_unit(unit)
+        {
+            uint32 maxmana = unit->GetMaxPower(POWER_MANA);
+            i_percent = maxmana ? unit->GetPower(POWER_MANA) * 100 / maxmana : 101;
+        }
+        Unit* getUnit() const { return i_unit; }
+        uint32 getPercent() const { return i_percent; }
+    private:
+        Unit* i_unit;
+        uint32 i_percent;
+};
+
+struct PrioritizeMana
+{
+    int operator()(PrioritizeManaUnitWraper const& x, PrioritizeManaUnitWraper const& y) const
+    {
+        return x.getPercent() > y.getPercent();
+    }
+};
+
+typedef std::priority_queue<PrioritizeManaUnitWraper, std::vector<PrioritizeManaUnitWraper>, PrioritizeMana> PrioritizeManaUnitQueue;
+
+class PrioritizeHealthUnitWraper
+{
+    public:
+        explicit PrioritizeHealthUnitWraper(Unit* unit) : i_unit(unit)
+        {
+            uint32 maxhealth = unit->GetMaxHealth();
+            i_percent = maxhealth ? unit->GetHealth() * 100 / maxhealth : 0;
+        }
+        Unit* getUnit() const { return i_unit; }
+        uint32 getPercent() const { return i_percent; }
+    private:
+        Unit* i_unit;
+        uint32 i_percent;
+};
+
+struct PrioritizeHealth
+{
+    int operator()(PrioritizeHealthUnitWraper const& x, PrioritizeHealthUnitWraper const& y) const
+    {
+        return x.getPercent() > y.getPercent();
+    }
+};
+
+typedef std::priority_queue<PrioritizeHealthUnitWraper, std::vector<PrioritizeHealthUnitWraper>, PrioritizeHealth> PrioritizeHealthUnitQueue;
+
+/**
+ * Fill target list by units around (x,y) points at radius distance
+
+ * @param targetUnitMap        Reference to target list that filled by function
+ * @param radius               Radius around (x,y) for target search
+ * @param pushType             Additional rules for target area selection (in front, angle, etc)
+ * @param spellTargets         Additional rules for target selection base at hostile/friendly state to original spell caster
+ * @param originalCaster       If provided set alternative original caster, if =NULL then used Spell::GetAffectiveObject() return
+ */
+void Spell::FillAreaTargets(UnitList& targetUnitMap, float radius, SpellNotifyPushType pushType, SpellTargets spellTargets, WorldObject* originalCaster /*=NULL*/)
+{
+    MaNGOS::SpellNotifierCreatureAndPlayer notifier(*this, targetUnitMap, radius, pushType, spellTargets, originalCaster);
+    Cell::VisitAllObjects(notifier.GetCenterX(), notifier.GetCenterY(), m_caster->GetMap(), notifier, radius);
+}
+
+/**
+ * @brief Fills a target list with party or raid members around a reference unit.
+ *
+ * @param targetUnitMap The target list being populated.
+ * @param member The reference member.
+ * @param radius The search radius.
+ * @param raid True to include the whole raid; false to limit to the subgroup.
+ * @param withPets True to include pets.
+ * @param withcaster True to include the caster when applicable.
+ */
+void Spell::FillRaidOrPartyTargets(UnitList& targetUnitMap, Unit* member, Unit* center, float radius, bool raid, bool withPets, bool withcaster)
+{
+    Player* pMember = member->GetCharmerOrOwnerPlayerOrPlayerItself();
+    Group* pGroup = pMember ? pMember->GetGroup() : NULL;
+
+    if (pGroup)
+    {
+        uint8 subgroup = pMember->GetSubGroup();
+
+        for (GroupReference* itr = pGroup->GetFirstMember(); itr != NULL; itr = itr->next())
+        {
+            Player* Target = itr->getSource();
+
+            // IsHostileTo check duel and controlled by enemy
+            if (Target && (raid || subgroup == Target->GetSubGroup())
+                && !m_caster->IsHostileTo(Target))
+            {
+                if ((Target == center || center->IsWithinDistInMap(Target, radius)) &&
+                        (withcaster || Target != m_caster))
+                {
+                    targetUnitMap.push_back(Target);
+                }
+
+                if (withPets)
+                    if (Pet* pet = Target->GetPet())
+                        if ((pet == center || center->IsWithinDistInMap(pet, radius)) &&
+                                (withcaster || pet != m_caster))
+                        {
+                            targetUnitMap.push_back(pet);
+                        }
+            }
+        }
+    }
+    else
+    {
+        Unit* ownerOrSelf = pMember ? pMember : member->GetCharmerOrOwnerOrSelf();
+        if ((ownerOrSelf == center || center->IsWithinDistInMap(ownerOrSelf, radius)) &&
+                (withcaster || ownerOrSelf != m_caster))
+        {
+            targetUnitMap.push_back(ownerOrSelf);
+        }
+
+        if (withPets)
+        {
+            if (Pet* pet = ownerOrSelf->GetPet())
+            {
+                if ((pet == center || center->IsWithinDistInMap(pet, radius)) &&
+                        (withcaster || pet != m_caster))
+                {
+                    targetUnitMap.push_back(pet);
+                }
+            }
+        }
+    }
+}
+
+void Spell::FillRaidOrPartyManaPriorityTargets(UnitList& targetUnitMap, Unit* member, Unit* center, float radius, uint32 count, bool raid, bool withPets, bool withCaster)
+{
+    FillRaidOrPartyTargets(targetUnitMap, member, center, radius, raid, withPets, withCaster);
+
+    PrioritizeManaUnitQueue manaUsers;
+    for (UnitList::const_iterator itr = targetUnitMap.begin(); itr != targetUnitMap.end(); ++itr)
+        if ((*itr)->GetPowerType() == POWER_MANA && !(*itr)->IsDead())
+        {
+            manaUsers.push(PrioritizeManaUnitWraper(*itr));
+        }
+
+    targetUnitMap.clear();
+    while (!manaUsers.empty() && targetUnitMap.size() < count)
+    {
+        targetUnitMap.push_back(manaUsers.top().getUnit());
+        manaUsers.pop();
+    }
+}
+
+void Spell::FillRaidOrPartyHealthPriorityTargets(UnitList& targetUnitMap, Unit* member, Unit* center, float radius, uint32 count, bool raid, bool withPets, bool withCaster)
+{
+    FillRaidOrPartyTargets(targetUnitMap, member, center, radius, raid, withPets, withCaster);
+
+    PrioritizeHealthUnitQueue healthQueue;
+    for (UnitList::const_iterator itr = targetUnitMap.begin(); itr != targetUnitMap.end(); ++itr)
+        if (!(*itr)->IsDead())
+        {
+            healthQueue.push(PrioritizeHealthUnitWraper(*itr));
+        }
+
+    targetUnitMap.clear();
+    while (!healthQueue.empty() && targetUnitMap.size() < count)
+    {
+        targetUnitMap.push_back(healthQueue.top().getUnit());
+        healthQueue.pop();
+    }
+}
